Only patch include directories on Linux

+++ recipe/build_clangdev.sh
@@ -1,9 +1,10 @@
+if [ "$(uname)" == "Linux" ]; then
+    sed -i.bak -e 's@addSystemInclude(DriverArgs, CC1Args, SysRoot + "/usr/local/include");@addSystemInclude(DriverArgs, CC1Args, "'"${PREFIX}/${HOST}/sysroot/usr/include"'");@g' \
+        lib/Driver/ToolChains/Linux.cpp && rm $_.bak
 
-sed -i.bak -e 's@addSystemInclude(DriverArgs, CC1Args, SysRoot + "/usr/local/include");@addSystemInclude(DriverArgs, CC1Args, "'"${PREFIX}/${HOST}/sysroot/usr/include"'");@g' \
-	    lib/Driver/ToolChains/Linux.cpp && rm $_.bak
-
-sed -i.bak -e 's@AddPath("/usr/local/include", System, false);@AddPath("'"${PREFIX}/${HOST}/sysroot/usr/include"'", System, false);@g' \
-	    lib/Frontend/InitHeaderSearch.cpp && rm $_.bak
+    sed -i.bak -e 's@AddPath("/usr/local/include", System, false);@AddPath("'"${PREFIX}/${HOST}/sysroot/usr/include"'", System, false);@g' \
+        lib/Frontend/InitHeaderSearch.cpp && rm $_.bak
+fi
 
 mkdir build
 cd build
